+++ IFI_Fresher_Test/src/main/java/com/ifi/fresher_test/ifi_fresher_test/service/ContestantService.java
@@ -1,11 +1,8 @@
 package com.ifi.fresher_test.ifi_fresher_test.service;
 
-import com.ifi.fresher_test.ifi_fresher_test.dto.AccountDTO;
 import com.ifi.fresher_test.ifi_fresher_test.dto.ContestantDTO;
 import com.ifi.fresher_test.ifi_fresher_test.mapper.ContestantMapper;
-import com.ifi.fresher_test.ifi_fresher_test.model.Account;
 import com.ifi.fresher_test.ifi_fresher_test.model.Contestant;
-import com.ifi.fresher_test.ifi_fresher_test.repository.AccountRepository;
 import com.ifi.fresher_test.ifi_fresher_test.repository.ContestantRepository;
 import com.ifi.fresher_test.ifi_fresher_test.util.MessageResource;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +11,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Optional;
 
@@ -39,8 +35,8 @@ public class ContestantService {
     }
 
     public ResponseEntity<ContestantDTO> findContestantByUsername(String username) {
-        Optional<Contestant> optionalAccount = contestantRepository.findById(username);
-        return optionalAccount.map(contestant -> new ResponseEntity<>(
+        Optional<Contestant> contestantOptional = contestantRepository.findById(username);
+        return contestantOptional.map(contestant -> new ResponseEntity<>(
                 ContestantMapper.entityToDTO(contestant), HttpStatus.OK)
         ).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
@@ -48,14 +44,18 @@ public class ContestantService {
     public ResponseEntity<?> addContestant(ContestantDTO contestantDTO) {
         try {
             if (!contestantRepository.findById(contestantDTO.getUsername()).isPresent()) {
-                Contestant contestant = ContestantMapper.dtoToEntity(contestantDTO);
-                contestantRepository.save(contestant);
-                return new ResponseEntity<ContestantDTO>(
-                        new ContestantDTO(
-                                contestant.getUsername(),
-                                contestant.getFullName()
-                        ), HttpStatus.CREATED
-                );
+                if (accountService.accountRepository.findById(contestantDTO.getUsername()).get().getRole().equals(MessageResource.CONTESTANT)) {
+                    Contestant contestant = ContestantMapper.dtoToEntity(contestantDTO);
+                    contestantRepository.save(contestant);
+                    return new ResponseEntity<ContestantDTO>(
+                            new ContestantDTO(
+                                    contestant.getUsername(),
+                                    contestant.getFullName()
+                            ), HttpStatus.CREATED
+                    );
+                } else {
+                    return new ResponseEntity<String>(MessageResource.CONTRIBUTOR_ALREADY_EXISTS, HttpStatus.ALREADY_REPORTED);
+                }
             } else {
                 return new ResponseEntity<String>(MessageResource.CONTESTANT_ALREADY_EXISTS, HttpStatus.ALREADY_REPORTED);
             }
@@ -64,27 +64,35 @@ public class ContestantService {
         }
     }
 
-    public ResponseEntity<ContestantDTO> updateContestant(String username, ContestantDTO contestantDTO) {
-        Optional<Contestant> contestantOptional = contestantRepository.findById(username);
-        return contestantOptional.map(contestant -> {
-            contestant.setFullName(contestantDTO.getFullName());
-            contestantRepository.save(contestant);
-            return new ResponseEntity<>(new ContestantDTO(
-                    contestant.getUsername(),
-                    contestant.getFullName()
-            ), HttpStatus.OK);
-        }).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    public ResponseEntity<?> updateContestant(String username, ContestantDTO contestantDTO) {
+        Optional<Contestant> optionalContestant = contestantRepository.findById(username);
+        if (optionalContestant.isPresent()) {
+            return optionalContestant.map(contestant -> {
+                contestant.setFullName(contestantDTO.getFullName());
+                contestantRepository.save(contestant);
+                return new ResponseEntity<ContestantDTO>(new ContestantDTO(
+                        contestant.getUsername(),
+                        contestant.getFullName()
+                ), HttpStatus.OK);
+            }).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        } else {
+            return new ResponseEntity<String>(MessageResource.ACCOUNT_NOT_CREATED_YET + " or " + MessageResource.CONTESTANT_NOT_CREATED_YET, HttpStatus.NOT_FOUND);
+        }
     }
 
-    public ResponseEntity<ContestantDTO> deleteContestant(String username) {
-        Optional<Contestant> consultantOptional = contestantRepository.findById(username);
-        return consultantOptional.map(contestant -> {
-            contestantRepository.delete(contestant);
-            accountService.deleteAccount(contestant.getUsername());
-            return new ResponseEntity<>(new ContestantDTO(
-                    contestant.getUsername(),
-                    contestant.getFullName()
-            ), HttpStatus.OK);
-        }).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+    public ResponseEntity<?> deleteContestant(String username) {
+        Optional<Contestant> optionalContestant = contestantRepository.findById(username);
+        if (optionalContestant.isPresent()) {
+            return optionalContestant.map(contestant -> {
+                contestantRepository.delete(contestant);
+                accountService.deleteAccount(contestant.getUsername());
+                return new ResponseEntity<ContestantDTO>(new ContestantDTO(
+                        contestant.getUsername(),
+                        contestant.getFullName()
+                ), HttpStatus.OK);
+            }).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        } else {
+            return new ResponseEntity<String>(MessageResource.ACCOUNT_NOT_CREATED_YET + " or " + MessageResource.CONTESTANT_NOT_CREATED_YET, HttpStatus.NOT_FOUND);
+        }
     }
 }
